Move tests into the proper module
