+++ packages/shader-composer/src/stdlib/casts.test.ts
@@ -1,7 +1,7 @@
 import { Color, Vector2, Vector3 } from "three"
 import { glslRepresentation } from "../glslRepresentation"
 import { float, vec3, vec4 } from "./casts"
-import { Int, NewVec3, Vec3 } from "./values"
+import { Int } from "./values"
 
 describe("float", () => {
   it("returns an expression that casts the given value to a float", () => {
@@ -41,39 +41,6 @@ describe("vec3", () => {
     const cast = vec4(new Vector2(1, 2), new Vector2(3, 4))
     expect(glslRepresentation(cast)).toEqual(
       `vec4(vec2(1.0, 2.0), vec2(3.0, 4.0))`
-    )
-  })
-})
-
-describe("NewVec3", () => {
-  it("value is a unit", () => {
-    const unit = Vec3(new Vector3(1, 2, 3))
-    const v = NewVec3(unit)
-    expect(glslRepresentation(v._unitConfig.value)).toEqual(
-      `vec3(${glslRepresentation(unit)})`
-    )
-  })
-
-  it("value is a JS value with its own GLSL representation", () => {
-    const value = new Vector3(1, 2, 3)
-    const v = NewVec3(value)
-    expect(glslRepresentation(v._unitConfig.value)).toEqual(
-      `vec3(vec3(1.0, 2.0, 3.0))`
-    )
-  })
-
-  it("value is an array of components", () => {
-    const value = [1, 2, 3]
-    const v = NewVec3(value)
-    expect(glslRepresentation(v._unitConfig.value)).toEqual(
-      `vec3(1.0, 2.0, 3.0)`
-    )
-  })
-
-  it("value is a 'nested' array. Well, kinda.", () => {
-    const v = NewVec3([1, ...[2, 3]])
-    expect(glslRepresentation(v._unitConfig.value)).toEqual(
-      `vec3(1.0, 2.0, 3.0)`
     )
   })
 })
